requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
